feat(project-creation): added hook to all callsites
